requirements.txt
#requirements.txt

fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0
pandas>=2.1.3
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0
scipy>=1.11.4
scikit-learn>=1.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
python-multipart>=0.0.6
Pillow>=10.1.0
lxml>=4.9.3
google-generativeai>=0.3.2
psutil>=5.9.6
